refactor: Refactored transforms
