+++ doctr/transforms/modules/__init__.py
@@ -1,3 +1,5 @@
+from .base import *
+
 from doctr.file_utils import is_tf_available, is_torch_available
 
 if is_tf_available():
